reporting-api/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
python-dotenv==1.0.1
flask-cors==5.0.0
psycopg2-binary==2.9.10
